refactor : #7 mainPage 수정: 반응형 레이아웃 수정
